修复BUG解决冲突保证后台框架能正常使用

+++ src/main/java/top/hellodays/blog_demo1/entity/Blog.java
@@ -8,8 +8,8 @@ import java.time.LocalDateTime;
 @Data
 public class Blog {
 
-    private Integer id;
-    private Integer user_id;
+    private Long id;
+    private Long user_id;
     @NotBlank(message = "标题不能为空")
     private String title;
     private String description;
